added authorization logic
