+++ src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/contract/ContractCancelDTO.java
@@ -1,4 +1,4 @@
-    package ch.uzh.ifi.hase.soprafs24.rest.dto;
+    package ch.uzh.ifi.hase.soprafs24.rest.dto.contract;
 
 import lombok.Getter;
 import lombok.Setter;
