+++ src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/contract/ContractGetDTO.java
@@ -1,8 +1,13 @@
-package ch.uzh.ifi.hase.soprafs24.rest.dto;
+package ch.uzh.ifi.hase.soprafs24.rest.dto.contract;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
-public class ContractPostDTO {
+import ch.uzh.ifi.hase.soprafs24.constant.ContractStatus;
+import ch.uzh.ifi.hase.soprafs24.rest.dto.LocationDTO;
+
+public class ContractGetDTO {
+    private Long contractId;
     private String title;
     private float mass;
     private float volume;
@@ -17,8 +22,20 @@ public class ContractPostDTO {
     private LocationDTO fromLocation;
     private LocationDTO toLocation;
     private LocalDateTime moveDateTime;
+    private ContractStatus contractStatus;
+    private LocalDateTime creationDateTime;
+    private List<String> contractPhotos;
+    private String cancelReason;
 
     // Getters and Setters
+    public Long getContractId() {
+        return contractId;
+    }
+
+    public void setContractId(Long contractId) {
+        this.contractId = contractId;
+    }
+
     public String getTitle() {
         return title;
     }
@@ -129,5 +146,37 @@ public class ContractPostDTO {
 
     public void setMoveDateTime(LocalDateTime moveDateTime) {
         this.moveDateTime = moveDateTime;
+    }
+
+    public ContractStatus getContractStatus() {
+        return contractStatus;
+    }
+
+    public void setContractStatus(ContractStatus contractStatus) {
+        this.contractStatus = contractStatus;
+    }
+
+    public LocalDateTime getCreationDateTime() {
+        return creationDateTime;
+    }
+
+    public void setCreationDateTime(LocalDateTime creationDateTime) {
+        this.creationDateTime = creationDateTime;
+    }
+
+    public List<String> getContractPhotos() {
+        return contractPhotos;
+    }
+
+    public void setContractPhotos(List<String> contractPhotos) {
+        this.contractPhotos = contractPhotos;
+    }
+
+    public String getCancelReason() {
+        return cancelReason;
+    }
+
+    public void setCancelReason(String cancelReason) {
+        this.cancelReason = cancelReason;
     }
 }
