+++ src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/offer/OfferGetDTO.java
@@ -1,9 +1,10 @@
-package ch.uzh.ifi.hase.soprafs24.rest.dto;
+package ch.uzh.ifi.hase.soprafs24.rest.dto.offer;
 
 import java.time.LocalDateTime;
 
 import ch.uzh.ifi.hase.soprafs24.constant.OfferStatus;
 import ch.uzh.ifi.hase.soprafs24.rest.dto.auth.response.AuthenticatedDriverDTO;
+import ch.uzh.ifi.hase.soprafs24.rest.dto.contract.ContractGetDTO;
 import lombok.Getter;
 import lombok.Setter;
