+++ src/main/java/ch/uzh/ifi/hase/soprafs24/rest/dto/offer/OfferPostDTO.java
@@ -1,4 +1,4 @@
-package ch.uzh.ifi.hase.soprafs24.rest.dto;
+package ch.uzh.ifi.hase.soprafs24.rest.dto.offer;
 
 import lombok.Getter;
 import lombok.Setter;
