+++ src/test/java/ch/uzh/ifi/hase/soprafs24/service/ContractPollingServiceTest.java
@@ -12,11 +12,11 @@ import org.junit.jupiter.api.Test;
 import org.mockito.*;
 
 import ch.uzh.ifi.hase.soprafs24.entity.Contract;
-import ch.uzh.ifi.hase.soprafs24.rest.dto.ContractGetDTO;
-import ch.uzh.ifi.hase.soprafs24.rest.dto.ContractFilterDTO;
 import ch.uzh.ifi.hase.soprafs24.service.ContractPollingService;
 import ch.uzh.ifi.hase.soprafs24.service.ContractService;
 import ch.uzh.ifi.hase.soprafs24.repository.ContractRepository;
+import ch.uzh.ifi.hase.soprafs24.rest.dto.contract.ContractFilterDTO;
+import ch.uzh.ifi.hase.soprafs24.rest.dto.contract.ContractGetDTO;
 
 class ContractPollingServiceTest {
